Pull some information out into the about page
